package: change usecase folders
