+++ app/src/main/java/com/eatssu/android/data/usecase/auth/GetRefreshTokenUseCase.kt
@@ -1,15 +1,15 @@
-package com.eatssu.android.data.usecase
+package com.eatssu.android.data.usecase.auth
 
 import android.content.Context
 import com.eatssu.android.util.MySharedPreferences
 import dagger.hilt.android.qualifiers.ApplicationContext
 import javax.inject.Inject
 
-class GetUserNameUseCase @Inject constructor(
-//    private val preferencesRepository: PreferencesRepository,
+class GetRefreshTokenUseCase @Inject constructor(
     @ApplicationContext private val context: Context
+//    private val preferencesRepository: PreferencesRepository,
 ) {
     suspend operator fun invoke(): String {
-        return MySharedPreferences.getUserName(context)
+        return MySharedPreferences.getRefreshToken(context)
     }
 }
