+++ app/src/main/java/com/eatssu/android/data/usecase/auth/GetUserEmailUseCase.kt
@@ -1,15 +1,15 @@
-package com.eatssu.android.data.usecase
+package com.eatssu.android.data.usecase.auth
 
 import android.content.Context
 import com.eatssu.android.util.MySharedPreferences
 import dagger.hilt.android.qualifiers.ApplicationContext
 import javax.inject.Inject
 
-class GetRefreshTokenUseCase @Inject constructor(
+class GetUserEmailUseCase @Inject constructor(
     @ApplicationContext private val context: Context
-//    private val preferencesRepository: PreferencesRepository,
 ) {
     suspend operator fun invoke(): String {
-        return MySharedPreferences.getRefreshToken(context)
+
+        return MySharedPreferences.getUserEmail(context)
     }
 }
