+++ app/src/main/java/com/eatssu/android/data/usecase/auth/GetUserNameUseCase.kt
@@ -1,15 +1,15 @@
-package com.eatssu.android.data.usecase
+package com.eatssu.android.data.usecase.auth
 
 import android.content.Context
 import com.eatssu.android.util.MySharedPreferences
 import dagger.hilt.android.qualifiers.ApplicationContext
 import javax.inject.Inject
 
-class GetUserEmailUseCase @Inject constructor(
+class GetUserNameUseCase @Inject constructor(
+//    private val preferencesRepository: PreferencesRepository,
     @ApplicationContext private val context: Context
 ) {
     suspend operator fun invoke(): String {
-
-        return MySharedPreferences.getUserEmail(context)
+        return MySharedPreferences.getUserName(context)
     }
 }
