+++ app/src/main/java/com/eatssu/android/data/usecase/auth/ReissueTokenUseCase.kt
@@ -1,15 +1,14 @@
-package com.eatssu.android.data.usecase
+package com.eatssu.android.data.usecase.auth
 
 import com.eatssu.android.base.BaseResponse
-import com.eatssu.android.data.dto.request.LoginWithKakaoRequest
 import com.eatssu.android.data.dto.response.TokenResponse
 import com.eatssu.android.data.repository.OauthRepository
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
-class LoginUseCase @Inject constructor(
+class ReissueTokenUseCase @Inject constructor(
     private val oauthRepository: OauthRepository,
 ) {
-    suspend operator fun invoke(body: LoginWithKakaoRequest): Flow<BaseResponse<TokenResponse>> =
-        oauthRepository.login(body)
+    suspend operator fun invoke(refreshToken: String): Flow<BaseResponse<TokenResponse>> =
+        oauthRepository.reissueToken(refreshToken)
 }
