+++ app/src/main/java/com/eatssu/android/data/usecase/auth/SetUserNameUseCase.kt
@@ -1,4 +1,4 @@
-package com.eatssu.android.data.usecase
+package com.eatssu.android.data.usecase.auth
 
 import android.content.Context
 import com.eatssu.android.base.BaseResponse
