+++ app/src/main/java/com/eatssu/android/data/usecase/menu/GetMenuNameListUseCase.kt
@@ -1,4 +1,4 @@
-package com.eatssu.android.data.usecase
+package com.eatssu.android.data.usecase.menu
 
 import com.eatssu.android.base.BaseResponse
 import com.eatssu.android.data.dto.response.GetMenuReviewInfoResponse
@@ -6,7 +6,7 @@ import com.eatssu.android.data.repository.ReviewRepository
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
-class GetMenuReviewInfoUseCase @Inject constructor(
+class GetMenuNameListUseCase @Inject constructor(
     private val reviewRepository: ReviewRepository,
 ) {
     suspend operator fun invoke(menuId: Long): Flow<BaseResponse<GetMenuReviewInfoResponse>> =
