+++ app/src/main/java/com/eatssu/android/data/usecase/review/GetMenuReviewListUseCase.kt
@@ -1,4 +1,4 @@
-package com.eatssu.android.data.usecase
+package com.eatssu.android.data.usecase.review
 
 import com.eatssu.android.base.BaseResponse
 import com.eatssu.android.data.dto.response.GetReviewListResponse
@@ -7,11 +7,11 @@ import com.eatssu.android.data.repository.ReviewRepository
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
-class GetMealReviewListUseCase @Inject constructor(
+class GetMenuReviewListUseCase @Inject constructor(
     private val reviewRepository: ReviewRepository,
 ) {
     suspend operator fun invoke(
-        mealId: Long?,
+        menuId: Long?,
     ): Flow<BaseResponse<GetReviewListResponse>> =
-        reviewRepository.getReviewList(MenuType.VARIABLE.toString(), mealId, 0)
+        reviewRepository.getReviewList(MenuType.FIXED.toString(), 0, menuId)
 }
