+++ app/src/main/java/com/eatssu/android/ui/main/MainViewModel.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.eatssu.android.R
-import com.eatssu.android.data.usecase.GetUserInfoUseCase
+import com.eatssu.android.data.usecase.auth.GetUserInfoUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.flow.MutableStateFlow
