+++ app/src/main/java/com/eatssu/android/ui/mypage/MyPageViewModel.kt
@@ -4,14 +4,14 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.eatssu.android.BuildConfig
 import com.eatssu.android.data.repository.PreferencesRepository
-import com.eatssu.android.data.usecase.AlarmUseCase
-import com.eatssu.android.data.usecase.GetDailyNotificationStatusUseCase
-import com.eatssu.android.data.usecase.GetUserInfoUseCase
-import com.eatssu.android.data.usecase.LogoutUseCase
-import com.eatssu.android.data.usecase.SetAccessTokenUseCase
-import com.eatssu.android.data.usecase.SetDailyNotificationStatusUseCase
-import com.eatssu.android.data.usecase.SetRefreshTokenUseCase
-import com.eatssu.android.data.usecase.SignOutUseCase
+import com.eatssu.android.data.usecase.alarm.AlarmUseCase
+import com.eatssu.android.data.usecase.alarm.GetDailyNotificationStatusUseCase
+import com.eatssu.android.data.usecase.alarm.SetDailyNotificationStatusUseCase
+import com.eatssu.android.data.usecase.auth.GetUserInfoUseCase
+import com.eatssu.android.data.usecase.auth.LogoutUseCase
+import com.eatssu.android.data.usecase.auth.SetAccessTokenUseCase
+import com.eatssu.android.data.usecase.auth.SetRefreshTokenUseCase
+import com.eatssu.android.data.usecase.auth.SignOutUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
