+++ app/src/main/java/com/eatssu/android/ui/mypage/myreview/MyReviewViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.eatssu.android.data.dto.response.toReviewList
 import com.eatssu.android.data.model.Review
-import com.eatssu.android.data.usecase.GetMyReviewsUseCase
+import com.eatssu.android.data.usecase.auth.GetMyReviewsUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
