+++ app/src/main/java/com/eatssu/android/ui/review/write/ReviewWriteViewModel.kt
@@ -3,7 +3,7 @@ package com.eatssu.android.ui.review.write
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.eatssu.android.data.dto.request.WriteReviewRequest
-import com.eatssu.android.data.usecase.WriteReviewUseCase
+import com.eatssu.android.data.usecase.review.WriteReviewUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
